delete
git-svn-id: https://thelia.svn.sourceforge.net/svnroot/thelia@823 6d4e0a53-8826-0410-8364-8b493f4a4454

+++ thelia/classes/Commande.class.php
@@ -78,6 +78,10 @@
 		}
 		
 		function supprimer(){
+
+            if($this->id == "")
+                    return 0;
+			
 			$venteprod = new Venteprod();
 			$query = "delete from $venteprod->table where commande='" . $this->id . "'";
 			$resul = mysql_query($query, $this->link);
